refactor(engine): promote trigger components to engine layer
